Refactor voice selection and preset assignment logic
- Updated LEDMatrix to use uint32_t for TOTAL_LEDS.
- Adjusted OLED display logic to use zero-based indexing for voice selection.
- Modified AS5600Manager to pair voices for base value selection.
- Refactored Sequencer to extract common initialization logic and improve GPIO pin handling.
- Updated SequencerDefs to increase MIN_GATE_LENGTH_TICKS for better timing.
- Enhanced UI event handling to ensure voice selection is required for preset assignment.
- Added voice selection state tracking in UIState.
- Implemented voice preset management functions in VoicePresets.
- Created a test harness for simulating button events related to voice selection and preset assignment.

+++ src/LEDMatrix/ledMatrix.h
@@ -18,7 +18,7 @@ public:
   static constexpr uint8_t WIDTH = LEDConstants::MATRIX_WIDTH;
   static constexpr uint8_t HEIGHT = LEDConstants::MATRIX_HEIGHT;
   static constexpr uint8_t DATA_PIN = LEDConstants::MATRIX_DATA_PIN;
-  static constexpr uint8_t TOTAL_LEDS = LEDConstants::MATRIX_TOTAL_LEDS;
+  static constexpr uint32_t TOTAL_LEDS = LEDConstants::MATRIX_TOTAL_LEDS;
 
   /**
    * @brief Constructor - initializes LED array to black
